FoodAPITranslator & FoodAPIInterface Overhaul
The translator and interface were overhauled so that any piece of information desired can be called from one generic method instead of several near-identical methods that specialize on one piece of information. Recipe.java was also slightly edited to be compatible with these changes. The output is identical to what it was before.