The Zip Code Revision
The UserProfile and UserDatabase classes have been revised to use zip codes instead of IP addresses and the Main class has been updated to prevent crashes.

IMPORTANT: The test.txt file will not be able to load the first time this new version is ran. This isn't going to break anything and the file will be formatted properly after a new version is saved.